Reorganise the repository.
