--- spmpy/img_processing.py
@@ -1,9 +1,0 @@
-"""
-Collection of functions for image processing
-
-"""
-
-class ImageProcessing():
-
-    def __init__(self):
-        pass
